Plot interactive scatterplot matrices

Plot interactive scatterplot matrices.

plotSMApp(data = data, xbins = 10)

Arguments

data
DATA FRAME | Read counts

xbins
INTEGER | Number of bins partitioning the range of the plot; default 10

Value

A Shiny application that shows a scatterplot matrix with hexagon bins and allows users to click on hexagon bins to determine how many genes they each contain. The user can download a file that contains the gene IDs that are located in the clicked hexagon bin.

Examples

# Example: Create interactive scatterplot matrix for first two treatment
# groups of data.

data(soybean_cn_sub)
soybean_cn_sub <- soybean_cn_sub[,1:7]
app <- plotSMApp(data=soybean_cn_sub)
if (interactive()) {
    shiny::runApp(app)
}