Plot interactive volcano plots

Plot interactive volcano plots.

plotVolcanoApp(data = data, dataMetrics = dataMetrics,
  option = "hexagon", pointColor = "orange")

Arguments

data
DATA FRAME | Read counts

dataMetrics
LIST | Differential expression metrics. This object must contain one column named "logFC" and one column named "PValue".

option
CHARACTER STRING ["hexagon" | "allPoints"] | The background of plot; default "hexagon"

pointColor
CHARACTER STRING | Color of overlaid points on scatterplot matrix; default "orange"

Value

A Shiny application that shows a volcano plot and allows users to overlay genes depending on two values, usually a statistical value (such as P-value) and a magnitude change value (such as log fold change). The user can download a file that contains the gene IDs that pass these thresholds.

Examples

# Example 1: Create interactive volcano plot of logged data using hexagon
# bins for the background.

data(soybean_cn_sub)
data(soybean_cn_sub_metrics)
app <- plotVolcanoApp(data = soybean_cn_sub,
    dataMetrics = soybean_cn_sub_metrics)
if (interactive()) {
    shiny::runApp(app)
}

# Example 2: Create interactive volcano plot of logged data using points for 
# the background.

app <- plotVolcanoApp(data = soybean_cn_sub,
    dataMetrics = soybean_cn_sub_metrics, option = "allPoints",
    pointColor = "magenta")
if (interactive()) {
    shiny::runApp(app)
}