Raw soybean leaves iron-metabolism data

This dataset contains raw RNA-sequencing read counts from a soybean dataset that compared leaves that were exposed to iron-rich (iron -postive) soil conditions versus leaves that were exposed to iron-poor (iron-negative) soil conditions. The data was collected 120 minutes after iron conditions were initiated.

Format

a RData instance, 1 row per gene

Details

Raw soybean leaves data

ID gene name

N.1 iron-negative condition replicate 1 raw read counts

N.2 iron-negative condition replicate 2 raw read counts

N.3 iron-negative condition replicate 3 raw read counts

P.1 iron-positive condition replicate 1 raw read counts

P.2 iron-positive condition replicate 2 raw read counts

P.3 iron-positive condition replicate 3 raw read counts