Raw soybean leaves iron-metabolism metrics

This data contains metrics for raw RNA-sequencing read counts from a soybean dataset that compared leaves that were exposed to iron-rich (iron-postive) soil conditions versus leaves that were exposed to iron-poor (iron-negative) soil conditions. The data was collected 120 minutes after iron conditions were initiated. The metrics include the log fold change and the p-values for all genes and all pairwise combinations of treatment groups.

Format

a RData instance, 1 list per treatment group combination and 1 row per gene

Details

Raw soybean leaves metrics

ID gene name

logFC log fold change

PValue p-value

See also